+++ velox/experimental/stateful/agg/GroupAggregator.h
@@ -55,7 +55,7 @@ class GroupAggregator : public exec::Operator, public KeyedProcessFunction {
 
   void open(StreamOperatorStateHandler* stateHandler) override;
 
-  RowVectorPtr processElements(uint32_t key, RowVectorPtr input) override;
+  RowVectorPtr processElements(int64_t key, RowVectorPtr input) override;
 
   void close() override;
